pmp_backend/requirements.txt
flask == 1.0.2
pip == 19.0.1
SQLAlchemy == 1.2.17
pyjwt == 1.7.1
flask_sqlalchemy == 2.3.2
flask_cors == 3.0.7
flask_socketio == 3.1.2
gevent == 1.4.0
gevent-websocket == 0.10.1
pyopenssl == 19.0.0
gunicorn == 19.9.0
pytest == 4.2.1
urllib3 == 1.24.1
requests == 2.21.0
eventlet
pymysql
mysqlclient == 1.3.14